added some folders and began adding documentation and type safety
