+++ backend/logic/machine-learning/game/game.py
@@ -1,10 +1,8 @@
 import chess
-from constants import START_FEN
-import chess
 import chess.pgn
-import io
-import chess.pgn
-from io import StringIO
+
+from utilities.constants import START_FEN
+
 
 class Game:
     def __init__(self, game_id: str, fen: str = START_FEN, moves: str = "", start: str = START_FEN, last_move: str = "", greedy: bool = False):
@@ -16,11 +14,6 @@ class Game:
         self.greedy = greedy  # Boolean flag for greedy mode
         self.board = chess.Board(fen)  # Chess board initialized from FEN
 
-    def update_last_move(self, move: str):
-        """Updates last move, adds it to history, and updates the board."""
-        self.last_move = move
-        self.moves += f" {move}"  # Append move to PGN history
-        self.board.push_san(move)  # Apply move to board
 
     def get_moves_pairs(self):
         """Returns a list of moves played."""
@@ -29,18 +22,6 @@ class Game:
     def get_fen(self):
         """Returns the current board state in FEN notation."""
         return self.board.fen()
-
-    def make_pgn(self):
-        """Generates a PGN string including the starting FEN."""
-        return f'[FEN "{self.start}"]\n\n{self.moves}'
-
-    def reset_game(self):
-        """Resets the game to its initial state."""
-        self.fen = self.start
-        self.moves = ""
-        self.last_move = ""
-        self.board = chess.Board(self.start)
-        
         
     
 
@@ -51,10 +32,7 @@ def get_moves_from_pgn(board):
     # Create a PGN exporter
     exporter = chess.pgn.StringExporter(headers=False, variations=False, comments=False)
     
-    # Get PGN string
     pgn = game.accept(exporter)
-
-    # Remove newline characters
     return pgn.replace("\n", " ").replace("\r", "")
 
 def make_update_payload(board: chess.Board, greedy: bool = False):
